+++ ESI.NET/Logic/DogmaLogic.cs
@@ -27,7 +27,7 @@ namespace ESI.NET.Logic
         /// </summary>
         /// <param name="attribute_id"></param>
         /// <returns></returns>
-        public async Task<EsiResponse<AttributeInfo>> Attribute(int attribute_id, EsiCallOptions options = null)
+        public async Task<EsiResponse<AttributeInfo>> Attribute(long attribute_id, EsiCallOptions options = null)
             => await Execute<AttributeInfo>(_client, _config, RequestSecurity.Public, HttpMethod.Get, "/dogma/attributes/{attribute_id}/",
                 replacements: new Dictionary<string, string>()
                 {
@@ -50,7 +50,7 @@ namespace ESI.NET.Logic
         /// </summary>
         /// <param name="effect_id"></param>
         /// <returns></returns>
-        public async Task<EsiResponse<EffectInfo>> Effect(int effect_id, EsiCallOptions options = null)
+        public async Task<EsiResponse<EffectInfo>> Effect(long effect_id, EsiCallOptions options = null)
             => await Execute<EffectInfo>(_client, _config, RequestSecurity.Public, HttpMethod.Get, "/dogma/effects/{effect_id}/",
                 replacements: new Dictionary<string, string>()
                 {
@@ -65,7 +65,7 @@ namespace ESI.NET.Logic
         /// <param name="type_id"></param>
         /// <param name="item_id"></param>
         /// <returns></returns>
-        public async Task<EsiResponse<DynamicItem>> DynamicItem(int type_id, long item_id, EsiCallOptions options = null)
+        public async Task<EsiResponse<DynamicItem>> DynamicItem(long type_id, long item_id, EsiCallOptions options = null)
             => await Execute<DynamicItem>(_client, _config, RequestSecurity.Public, HttpMethod.Get, "/dogma/dynamic/items/{type_id}/{item_id}/",
                 replacements: new Dictionary<string, string>()
                 {
