+++ ESI.NET/Logic/PlanetaryInteractionLogic.cs
@@ -34,7 +34,7 @@ namespace ESI.NET.Logic
         /// </summary>
         /// <param name="planet_id"></param>
         /// <returns></returns>
-        public async Task<EsiResponse<ColonyLayout>> ColonyLayout(int planet_id, EsiCallOptions options)
+        public async Task<EsiResponse<ColonyLayout>> ColonyLayout(long planet_id, EsiCallOptions options)
             => await Execute<ColonyLayout>(_client, _config, RequestSecurity.Authenticated, HttpMethod.Get, "/characters/{character_id}/planets/{planet_id}/",
                 replacements: new Dictionary<string, string>()
                 {
@@ -60,7 +60,7 @@ namespace ESI.NET.Logic
         /// </summary>
         /// <param name="schematic_id"></param>
         /// <returns></returns>
-        public async Task<EsiResponse<Schematic>> SchematicInformation(int schematic_id, EsiCallOptions options = null)
+        public async Task<EsiResponse<Schematic>> SchematicInformation(long schematic_id, EsiCallOptions options = null)
             => await Execute<Schematic>(_client, _config, RequestSecurity.Public, HttpMethod.Get, "/universe/schematics/{schematic_id}/",
                 replacements: new Dictionary<string, string>()
                 {
